Fix invoice Penjualan dan detail invoice

+++ application/controllers/checkout.php
@@ -13,9 +13,13 @@ class Checkout extends CI_Controller
 
 	public function order_process()
 	{
-		$this->cart->destroy();
-        $this->load->view('templates/header');
-        $this->load->view('checkout/order_process');
-
+		$is_processed = $this->model_invoice->index();
+		if( $is_processed ){
+			$this->cart->destroy();
+			$this->load->view('templates/header');
+			$this->load->view('checkout/order_process');
+		}else{
+			echo "Maaf, Pesanan anda gagal diproses!!";
+		}
 	}
 }
